Fix bug on systems w/ mult. devs/platforms w/ different OCL versions
If cf4ocl was compiled for OCL >= 1.2, and any of the platforms
available on the system was OCL <= 1.1, a seg. fault occurred
when trying to release devices associated with the later, because
the platform doesn't support clReleaseDevice, which was introduced
for device fission purposes in OCL 1.2.

Now the clReleaseDevice function is only used if the device OpenCL
C version is >= 1.2. Technically, it should be the platform OpenCL
version to be checked, because its the platform that supplies the
clReleaseDevice function, not the device itself. Checking the
platform version at this stage would involve instantiating a
platform wrapper object just for the purpose of getting the OpenCL
version, which seems a bit wasteful. This should be improved
when possible.

In fixing this problem, a new ccl_device_get_opencl_c_version was
added to the library, which is probably useful for runtime
generation of kernels.