Better fix for bug described in commit 9aab638
More specifically, the platform OpenCL version is checked instead
of the device OpenCL C version through the CL_DEVICE_VERSION
parameter (which yields the same information as the
CL_PLATFORM_VERSION parameter).

Add the ccl_device_get_opencl_version() function, which returns
this value as an integer (100, 110, 120, 200, etc). Add small
test for this function.

Better specify in the documentation of the several
ccl_*_get_opencl_version() functions that the return value is an
integer like 100 for OpenCL 1.0, 110 for OpenCL 1.1, etc.

Release context in sub-devices test when such test is not possible.