Make cf4ocl API stable (buffer, program) indep. of OCL ver. (wip)
In other words, make cf4ocl functions always visible even if the
underlying OpenCL requirements are not met. If such functions are
invoked without support an error is thrown, but this avoids breaking
client code when compiled with different OpenCL versions.

Current changes are for buffer and program modules, more to come.