ccl_c: remove compile with incorrectly specified headers test because some OpenCL implementations are "smart" enough to find the correct headers anyway.
Additionally, in the compile with source file and erroneous header
test, we now specify the headers and their names directly instead
of passing the inc. folder as a compiler option "-I /inc/folder",
because otherwise some OpenCL implementations cache the source
kernels with the same name (but different content) and incorrectly
use them.